[OpenCL Textures] Fix memory management in texture pool
Previously, the size of the memory which should be allocated was
calculated as multiplication width on height. It doesn't work well in
case when one texture has big size in height and the next one big size
in width. We tried to reuse the allocated memory and every time when
the next texture with big size was used we reallocated the previous
one. It has huge impact on the performance.
Now we check two dimensions independently. So, in this case we will
check both dimensions and it helps us to avoid the situation with
cyclic memory reallocation.